Move information page handler to new package
